Export RectangleSelection and CircleSelection with withContainerSize based on children
