+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/CircleSelection/index.js
@@ -1,10 +1,6 @@
 // @flow
 import CircleSelection from './CircleSelection';
-import CircleSelectionRenderer from './CircleSelectionRenderer';
 import type {SelectionData} from './types';
 
 export default CircleSelection;
-export {
-    CircleSelectionRenderer,
-};
 export type {SelectionData};
