+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/CircleSelection/tests/CircleSelection.test.js
@@ -1,8 +1,7 @@
 // @flow
 import {mount} from 'enzyme';
 import React from 'react';
-import {CircleSelection} from '../CircleSelection';
-import CircleSelectionRenderer from '../CircleSelectionRenderer';
+import {CircleSelectionContainer} from '../CircleSelection';
 
 jest.mock('../../../utils/Translator', () => ({
     translate: jest.fn((key) => key),
@@ -13,14 +12,14 @@ jest.mock('../../../utils/DOM/afterElementsRendered');
 
 test('The component should render', () => {
     const view = mount(
-        <CircleSelection
+        <CircleSelectionContainer
             containerHeight={1000}
             containerWidth={2000}
             onChange={jest.fn()}
             value={{radius: 1, top: 3, left: 4}}
         >
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
     expect(view.render()).toMatchSnapshot();
@@ -30,9 +29,9 @@ test('The component should center and maximize the selection if no value is give
     const changeSpy = jest.fn();
 
     mount(
-        <CircleSelection containerHeight={1000} containerWidth={2000} onChange={changeSpy} value={undefined}>
+        <CircleSelectionContainer containerHeight={1000} containerWidth={2000} onChange={changeSpy} value={undefined}>
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
     expect(changeSpy).toBeCalledWith({left: 1000, radius: 500, top: 500});
@@ -42,7 +41,7 @@ test('The component should reset the value if modifiable circle is doubleclicked
     const changeSpy = jest.fn();
 
     const view = mount(
-        <CircleSelection
+        <CircleSelectionContainer
             containerHeight={1000}
             containerWidth={2000}
             minHeight={100}
@@ -51,7 +50,7 @@ test('The component should reset the value if modifiable circle is doubleclicked
             value={{radius: 100, top: 30, left: 40}}
         >
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
     view.find('ModifiableCircle').prop('onDoubleClick')();
@@ -63,7 +62,7 @@ test('The component should center and maximize the selection when a minRadius an
     const changeSpy = jest.fn();
 
     mount(
-        <CircleSelection
+        <CircleSelectionContainer
             containerHeight={1000}
             containerWidth={2000}
             maxRadius={200}
@@ -72,7 +71,7 @@ test('The component should center and maximize the selection when a minRadius an
             value={undefined}
         >
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
     expect(changeSpy).toBeCalledWith({left: 1000, radius: 200, top: 500});
@@ -82,17 +81,17 @@ test('The component should not allow the selection to move over the borders', ()
     const changeSpy = jest.fn();
 
     const view = mount(
-        <CircleSelection
+        <CircleSelectionContainer
             containerHeight={1000}
             containerWidth={2000}
             onChange={changeSpy}
             value={{left: 0, top: 0, radius: 2000}}
         >
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
-    view.find(CircleSelectionRenderer).first().children().first().instance().handleCircleChange(
+    view.find('CircleSelectionComponent').first().instance().handleCircleChange(
         {radius: 0, left: -10, top: -20}
     );
     expect(changeSpy).toBeCalledWith({radius: 0, top: 0, left: 0});
@@ -102,17 +101,17 @@ test('The component should not allow the selection to be bigger than the contain
     const changeSpy = jest.fn();
 
     const view = mount(
-        <CircleSelection
+        <CircleSelectionContainer
             containerHeight={1000}
             containerWidth={2000}
             onChange={changeSpy}
             value={{left: 0, top: 0, radius: 2000}}
         >
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
-    view.find(CircleSelectionRenderer).first().children().first().instance().handleCircleChange(
+    view.find('CircleSelectionComponent').first().instance().handleCircleChange(
         {radius: 5000, left: 0, top: 0}
     );
     expect(changeSpy).toBeCalledWith({radius: 2236, top: 0, left: 0});
@@ -122,7 +121,7 @@ test('The component should not round if told by the properties', () => {
     const changeSpy = jest.fn();
 
     const view = mount(
-        <CircleSelection
+        <CircleSelectionContainer
             containerHeight={1000}
             containerWidth={2000}
             minRadius={5}
@@ -131,7 +130,7 @@ test('The component should not round if told by the properties', () => {
             value={undefined}
         >
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
     expect(view.find('ModifiableCircle').prop('top')).toBeGreaterThan(166);
@@ -143,7 +142,7 @@ test('The component should work with percentage values if told by the properties
     const changeSpy = jest.fn();
 
     mount(
-        <CircleSelection
+        <CircleSelectionContainer
             containerHeight={1000}
             containerWidth={2000}
             onChange={changeSpy}
@@ -151,7 +150,7 @@ test('The component should work with percentage values if told by the properties
             value={undefined}
         >
             <p>Lorem ipsum</p>
-        </CircleSelection>
+        </CircleSelectionContainer>
     );
 
     expect(changeSpy).toBeCalledWith({top: 0.5, left: 0.5, radius: 0.25});
