+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/RectangleSelection/index.js
@@ -1,10 +1,6 @@
 // @flow
 import RectangleSelection from './RectangleSelection';
-import RectangleSelectionRenderer from './RectangleSelectionRenderer';
 import type {SelectionData} from './types';
 
 export default RectangleSelection;
-export {
-    RectangleSelectionRenderer,
-};
 export type {SelectionData};
